% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{capitalizeKeywords}
\alias{capitalizeKeywords}
\title{Capitalize Keywords in CVIR expressions}
\usage{
capitalizeKeywords(x)
}
\arguments{
\item{x}{character CVIR expressions}
}
\value{
character
}
\description{
Capitalize Keywords in CVIR expressions
}
