% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{cleanCVIR}
\alias{cleanCVIR}
\title{Clean a vector of CVIR expressions}
\usage{
cleanCVIR(cvir)
}
\arguments{
\item{cvir}{character vector of CVIR expressions}
}
\value{
character
}
\description{
Removes comments and tabs, shortens white space, removes zero-character lines, and combines lines that do not end with `"."`
}
