% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{CVIR_PUNCTUATION}
\alias{CVIR_PUNCTUATION}
\title{CVIR Punctuation}
\format{
An object of class \code{character} of length 27.
}
\usage{
data(CVIR_PUNCTUATION)
}
\description{
Lookup table for CVIR Punctuation (syntax where `CVIR_TOKEN` label does not occur in code)
}
\references{
GOV.USDA.NRCS.NASIS.Cvir.Antlr.CvirTokenTypes
}
\keyword{datasets}
