% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{CVIR_TOKEN}
\alias{CVIR_TOKEN}
\title{CVIR Tokens}
\format{
An object of class \code{numeric} of length 320.
}
\usage{
data(CVIR_TOKEN)
}
\description{
Lookup table for CVIR Tokens
}
\references{
GOV.USDA.NRCS.NASIS.Cvir.Antlr.CvirTokenTypes
}
\keyword{datasets}
