% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvirrr-package.R
\docType{package}
\name{cvirrr-package}
\alias{cvirrr}
\alias{cvirrr-package}
\title{cvirrr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
