% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{CVIRScript.character}
\alias{CVIRScript.character}
\alias{CVIRScript}
\title{Initiate a CVIRScript from character string}
\usage{
\method{CVIRScript}{character}(x)

CVIRScript(x)
}
\arguments{
\item{x}{a character vector containing CVIR code}
}
\description{
Initiate a CVIRScript from character string
}
