% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{parseCVIR.CVIRScript}
\alias{parseCVIR.CVIRScript}
\alias{parseCVIR}
\title{Parse a CVIR Script}
\usage{
\method{parseCVIR}{CVIRScript}(x)

parseCVIR(x)
}
\arguments{
\item{x}{a CVIRScript object}
}
\description{
Parse a CVIR Script
}
