% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{pop.CVIRScript}
\alias{pop.CVIRScript}
\alias{pop}
\title{"pop" data off the CVIRScript stack}
\usage{
\method{pop}{CVIRScript}(x, instruction, n = 1, what = "value")

pop(x, instruction, n = 1, what = "value")
}
\arguments{
\item{x}{a CVIRScript object}

\item{instruction}{one of: `"NUMERIC"`, `"VARIABLE"`, `"STRING_LITERAL"` or their bytecode equivalents (`84`, `97`, `220`)}

\item{n}{number of items to pop off the stack}

\item{what}{values to pop off the stack (default `"value"`)}
}
\description{
"pop" data off the CVIRScript stack
}
