% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process.CVIRScript}
\alias{process.CVIRScript}
\alias{process}
\title{Process a CVIR Script}
\usage{
\method{process}{CVIRScript}(x)

process(x)
}
\arguments{
\item{x}{a CVIRScript object}
}
\description{
Process a CVIR Script
}
