% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R, R/tokens.R
\name{is_keyword}
\alias{is_keyword}
\alias{token_to_id}
\alias{id_to_token}
\title{Convert CVIR Tokens IDs, Labels and Syntax}
\usage{
is_keyword(token)

token_to_id(token)

id_to_token(id, convert_punctuation = TRUE)
}
\arguments{
\item{token}{Token Label}

\item{id}{Internal Token Byte Value}

\item{convert_punctuation}{For `id_to_token()`: Convert Token ID to user-level syntax (`TRUE`) or internal label (`FALSE`); Default: `TRUE`}
}
\value{
a numeric vector of token IDs (`token_to_id()`), or a character vector of token labels or syntax (`id_to_token()`)
}
\description{
Convert CVIR Tokens IDs, Labels and Syntax
}
\examples{

token_to_id("DERIVE")

id_to_token(317)
}
