% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvirrr-package.R
\docType{package}
\name{cvirrr-package}
\alias{cvirrr}
\alias{cvirrr-package}
\title{cvirrr: NASIS Calculation Validation Interpretation and Report Parser}
\description{
Tools for evaluating Calculation Validation Interpretation Report (CVIR) scripts outside of the U.S. Department of Agriculture Natural Resources Conservation Service (USDA-NRCS) National Soil Information System (NASIS)
}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov}

}
\keyword{internal}
